resolved sampler2d array indexing

+++ examples/assets/gfx/opengl/5_opengl_flappy_bird/vertex.glsl
@@ -2,10 +2,10 @@
 
 layout (location = 0) in vec2 a_Position;
 layout (location = 1) in vec2 a_TexCoords;
-layout (location = 2) in int a_TextureUnit;
+layout (location = 2) in float a_TextureUnit;
 
 layout (location = 0) out vec2 v_TexCoords;
-layout (location = 1) out flat int v_TextureUnit;
+layout (location = 1) out flat float v_TextureUnit;
 
 uniform mat4 u_Model;
 uniform mat4 u_Projection;
